Refaktorert komponentene som var i bruk for å få bedre prosjektstruktur. Byttet over til å benytte CSS-modules på alle komponenter, slik at microfrontendene ikke overskriver hverandre når de hentes inn i tms-min-side. Litt kodeopprydding.
