+++ src/App.jsx
@@ -1,4 +1,6 @@
-import MainPage from "./components/mainPage/MainPage";
+import Sidetittel from "./components/sidetittel/Sidetittel";
+import VarslerList from "./components/varsler/varsler-list/VarslerList";
+import InngangVarslinger from "./components/inngang-varslinger/InngangVarslinger";
 import "@navikt/ds-css";
 import "./App.css";
 
@@ -6,7 +8,13 @@ function App() {
   return (
     <main className="main">
       <div className="app">
-        <MainPage />
+        <section className="page-wrapper-microfrontend">
+          <Sidetittel />
+          <section className="brukernotifikasjons-liste">
+            <VarslerList />
+          </section>
+          <InngangVarslinger />
+        </section>
       </div>
     </main>
   );
