--- src/components/arkiverknapp/ArkiverKnapp.css
@@ -1,14 +1,0 @@
-.beskjed-arkiver-knapp {
-    color: #0067C5;
-    font-weight: bold;
-    font-size: 1rem;
-    letter-spacing: 0.02rem;
-    border-radius: 2px;
-    padding: 0.3rem 0.5rem 0.3rem 0.5rem;
-}
-
-.beskjed-arkiver-knapp:hover {
-    cursor: pointer;
-    color: white;
-    background-color: #0067C5;
-}
