+++ src/components/arkiverknapp/ArkiverKnapp.module.css
@@ -1,0 +1,17 @@
+.btn {
+  color: #0067c5;
+  font-weight: bold;
+  font-size: 1rem;
+  letter-spacing: 0.02rem;
+  border-radius: 2px;
+  padding: 0.3rem 0.5rem 0.3rem 0.5rem;
+  border: none;
+  background-color: transparent;
+  height: 30px;
+}
+
+.btn:hover {
+  cursor: pointer;
+  color: white;
+  background-color: #0067c5;
+}
