--- src/components/mainPage/MainPage.css
@@ -1,45 +1,0 @@
-.page-wrapper-microfrontend {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  color: #272727;
-}
-
-.heading-wrapper {
-  justify-content: flex-start;
-}
-
-.brukernotifikasjons-liste {
-  margin: 0 0 0 0;
-}
-
-.brukernotifikasjons-liste > a {
-  margin-bottom: 1rem;
-}
-
-.lenkepanel-liten-wrapper {
-  display: flex;
-}
-
-.lenkepanel-stor-wrapper {
-  display: flex;
-  flex-direction: column;
-}
-
-.navds-link-panel__chevron {
-  width: 18px;
-  height: 18px;
-}
-
-
-@media (min-width: 0px) {
-  .lenkepanel-liten-wrapper {
-    flex-direction: column;
-  }
-}
-
-@media (min-width: 648px) {
-  .lenkepanel-liten-wrapper {
-    flex-direction: row;
-  }
-}
