--- src/components/notifikasjoner/notifikasjonsBoks/NotifikasjonsBoks.css
@@ -1,195 +1,0 @@
-.oppgave-wrapper {
-  border-radius: 8px;
-  padding: 1.5rem !important;
-}
-
-.oppgave-wrapper-selected {
-  background-color: #FFD799 !important;
-  border: 1px solid transparent !important;
-  padding: 1.5rem !important;
-}
-
-.beskjed-wrapper {
-  border-radius: 8px;
-  padding: 1.5rem !important;
-}
-
-.beskjed-wrapper-selected {
-  background-color: #99C4DD !important;
-  border: 1px solid transparent !important;
-  padding: 1.5rem !important;
-}
-
-.brukernotifikasjon-tekst-wrapper {
-  padding: 0 1rem 0 0;
-  justify-items: center;
-  align-items: center;
-}
-
-.brukernotifikasjon-tekst {
-  font-size: 18px !important;
-  font-weight: bolder !important;
-  margin-bottom: 0 !important;
-  text-decoration: none !important;
-  color: #272727 !important;
-  line-height: normal !important;
-}
-
-.brukernotifikasjon-dato {
-  font-size: 16px !important;
-  margin: 0 0 0 0 !important;
-  color: #707070;
-}
-
-.beskjed-arkiver {
-  display: flex;
-  margin-bottom: 2px;
-  padding: 1.5rem !important;
-}
-
-.beskjed-arkiver-content {
-  display: flex;
-  justify-content: flex-start;
-}
-
-.oppgave-ikon-wrapper {
-  display: flex;
-  align-self: center;
-  overflow: visible;
-  padding: 1rem;
-  margin: 0 -1rem 0 0.5rem;
-  background-color: #FFECCC;
-  border-radius: 50%;
-  color: #262626;
-}
-
-.oppgave-ikon-wrapper-inverted {
-  display: flex;
-  align-self: center;
-  overflow: visible;
-  padding: 1rem;
-  margin: 0 -1rem 0 0.5rem;
-  background-color: #FFECCC;
-  border-radius: 50%;
-  color: #262626;
-}
-
-.beskjed-ikon-wrapper {
-  display: flex;
-  align-self: center;
-  overflow: visible;
-  padding: 1rem;
-  margin: 0 -1rem 0 0.5rem;
-  background-color: #CCE2F0;
-  border-radius: 50%;
-  color: #262626;
-}
-
-.beskjed-ikon-wrapper-inverted {
-  display: flex;
-  align-self: center;
-  overflow: visible;
-  padding: 1rem;
-  margin: 0 -1rem 0 0.5rem;
-  background-color: #CCE2F0;
-  border-radius: 50%;
-  color: #262626;
-}
-
-.beskjed-arkiver-tekst-wrapper {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  margin: 0 2rem 0 2rem;
-}
-
-.beskjed-arkiver-tekst-wrapper > h2{
-  font-size: 18px !important;
-  margin: 0 0 0 0 !important;
-}
-
-.beskjed-arkiver-tekst-wrapper > p{
-  font-size: 16px !important;
-}
-
-.beskjed-arkiver-ikon {
-  display: flex;
-  align-self: center;
-  overflow: visible;
-  padding: 1rem;
-  margin: 0 -1rem 0 0.5rem;
-  background-color: #CCE2F0;
-  border-radius: 50%;
-  color: #262626;
-}
-
-.beskjed-arkiver-ikon-inverted {
-  display: flex;
-  align-self: center;
-  overflow: visible;
-  padding: 1rem;
-  margin: 0 -1rem 0 0.5rem;
-  background-color: #CCE2F0;
-  border-radius: 50%;
-  color: #262626;
-}
-
-.beskjed-arkiver-knapp {
-  display: flex;
-  align-self: center;
-  justify-self: right;
-}
-
-@media (min-width: 0px) {
-  .oppgave-wrapper {
-    width: 100%;
-    margin: 0 0.5rem 1rem 0.5rem;
-  }
-
-  .oppgave-wrapper-selected {
-    width: 100%;
-    margin: 0 0.5rem 1rem 0.5rem;
-  }
-  
-  .beskjed-wrapper {
-    width: 100%;
-    margin: 0 0.5rem 1rem 0.5rem;
-  }
-
-  .beskjed-wrapper-selected {
-    width: 100%;
-    margin: 0 0.5rem 1rem 0.5rem;
-  }
-
-  .beskjed-arkiver {
-    width: 100%;
-    margin: 0 0.5rem 1rem 0.5rem;
-  }
-}
-
-@media (min-width: 648px) {
-  .oppgave-wrapper {
-    width: 37.5rem;
-    margin: 0 0 1rem 0;
-  }
-
-  .oppgave-wrapper-selected {
-    width: 37.5rem;
-    margin: 0 0 1rem 0;
-  }
-  
-  .beskjed-wrapper {
-    width: 37.5rem;
-    margin: 0 0 1rem 0;
-  }
-
-  .beskjed-wrapper-selected {
-    width: 37.5rem;
-    margin: 0 0 1rem 0;
-  }
-
-  .beskjed-arkiver {
-    width: 37.5rem;
-    margin: 0 0 1rem 0;
-  }
-}
