+++ src/components/sidetittel/Sidetittel.module.css
@@ -1,21 +1,21 @@
-.heading-wrapper {
+.wrapper {
   display: flex;
   justify-content: flex-start;
 }
 
-.person-tekst {
+.tekst {
   text-transform: capitalize;
   margin: 0 !important;
 }
 
 @media (min-width: 0px) {
-  .heading-wrapper {
+  .wrapper {
     margin: 1rem 0 2rem 0.5rem;
   }
 }
 
 @media (min-width: 648px) {
-  .heading-wrapper {
+  .wrapper {
     margin: 1rem 0 2rem 0.5rem;
   }
 }
