+++ src/components/varsler/arkiverbar-beskjed/ArkiverbarBeskjed.module.css
@@ -1,0 +1,58 @@
+.beskjed {
+  display: flex;
+  margin-bottom: 2px;
+  padding: 1.5rem !important;
+  align-items: center;
+}
+
+.content {
+  display: flex;
+  justify-content: flex-start;
+}
+
+.tekst_wrapper {
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  margin: 0 2rem 0 2rem;
+}
+
+.tekst_wrapper > h2 {
+  font-size: 18px !important;
+  margin: 0 0 0 0 !important;
+}
+
+.tekst_wrapper > p {
+  font-size: 16px !important;
+}
+
+.ikon {
+  display: flex;
+  align-self: center;
+  overflow: visible;
+  padding: 1rem;
+  margin: 0 -1rem 0 0.5rem;
+  background-color: #cce2f0;
+  border-radius: 50%;
+  color: #262626;
+}
+
+.knapp {
+  display: flex;
+  align-self: center;
+  justify-self: center;
+}
+
+@media (min-width: 0px) {
+  .beskjed {
+    width: 100%;
+    margin: 0 0.5rem 1rem 0.5rem;
+  }
+}
+
+@media (min-width: 648px) {
+  .beskjed {
+    width: 37.5rem;
+    margin: 0 0 1rem 0;
+  }
+}
